backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.6.0
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
mistralai==1.7.0
google-generativeai>=0.3.0
openai==1.78.1
sqlalchemy>=2.0.23
aiosqlite>=0.17.0 # Ajouté pour le support async de SQLite
chardet>=5.0.0 # Ajouté pour la détection d'encodage
python-multipart>=0.0.7 # NOUVEAU: Requis pour l'upload de fichiers FastAPI

# Text Extraction libraries
pypdf>=4.0.0 # Ajouté pour l'extraction PDF
odfpy>=1.4.0 # Ajouté pour l'extraction ODT
beautifulsoup4>=4.12.0 # Ajouté pour l'extraction de texte brut depuis HTML

# Export libraries
python-docx>=1.1.0
xhtml2pdf>=0.2.11
EbookLib>=0.18
html2text>=2020.1.16 # Pour la conversion HTML -> Texte simple

# NLP library
spacy>=3.7.0 # Ajout de spaCy
# Modèle français (à télécharger séparément via CLI)
# python -m spacy download fr_core_news_md

# Test dependencies
pytest>=7.0.0
httpx==0.28.1 # Version mise à jour pour compatibilité avec mistralai 1.7.0, tests à vérifier
pytest-asyncio>=0.21.0 # Ajouté pour les tests async
